requirements.txt
streamlit
numpy
python-dotenv
langchain
openai
langchain_openai
langchain-community
langchain-core